refactor: renamed component to InfoBoxWrapper
