+++ src/frontend/src/lib/components/info/InfoBoxWrapper.svelte
